+++ packages/std/src/iterator.rs
@@ -4,11 +4,6 @@ use std::convert::TryFrom;
 /// A Key-Value pair, returned from our iterators
 pub type Pair<V = Vec<u8>> = (Vec<u8>, V);
 
-/// KV is a Key-Value pair, returned from our iterators
-#[deprecated(since = "0.14.0", note = "Renamed to Pair")]
-#[allow(clippy::upper_case_acronyms)]
-pub type KV<V = Vec<u8>> = Pair<V>;
-
 #[derive(Copy, Clone)]
 // We assign these to integers to provide a stable API for passing over FFI (to wasm and Go)
 pub enum Order {