+++ packages/std/src/lib.rs
@@ -38,8 +38,7 @@ pub use crate::ibc::{
     IbcPacketReceiveMsg, IbcPacketTimeoutMsg, IbcReceiveResponse, IbcTimeout, IbcTimeoutBlock,
 };
 #[cfg(feature = "iterator")]
-#[allow(deprecated)]
-pub use crate::iterator::{Order, Pair, KV};
+pub use crate::iterator::{Order, Pair};
 pub use crate::math::{Decimal, Decimal256, Fraction, Uint128, Uint256, Uint512, Uint64};
 pub use crate::query::{
     AllBalanceResponse, BalanceResponse, BankQuery, CustomQuery, QueryRequest, WasmQuery,